+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 2.0b2 (unreleased)
 ------------------
 
+- Package is now compatible with Plone 5.
+  [hvelarde]
+
 - Refactor logo validation to ensure image is in PNG format.
   [hvelarde]