+++ CHANGES.rst
@@ -4,7 +4,11 @@ Changelog
 2.0b2 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Package is now compatible with Plone 5.
+  [hvelarde]
+
+- Refactor logo validation to ensure image is in PNG format.
+  [hvelarde]
 
 
 2.0b1 (2015-12-01)